+++ Content.Client/IgnitionSource/IgnitionSourceSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared.IgnitionSource;
+
+namespace Content.Client.IgnitionSource;
+
+public sealed partial class IgnitionSourceSystem : SharedIgnitionSourceSystem;
