+++ Content.Server/IgnitionSource/IgniteOnTriggerComponent.cs
@@ -5,7 +5,7 @@ namespace Content.Server.IgnitionSource;
 
 /// <summary>
 /// Ignites for a certain length of time when triggered.
-/// Requires <see cref="IgnitionSourceComponent"/> along with triggering components.
+/// Requires <see cref="Shared.IgnitionSourceComponent"/> along with triggering components.
 /// </summary>
 [RegisterComponent, Access(typeof(IgniteOnTriggerSystem))]
 public sealed partial class IgniteOnTriggerComponent : Component
