+++ Content.Shared/IgnitionSource/IgnitionEvent.cs
@@ -1,7 +1,7 @@
 namespace Content.Shared.IgnitionSource;
 
 /// <summary>
-///     Raised in order to toggle the ignitionSourceComponent on an entity on or off
+///     Raised in order to toggle the <see cref="IgnitionSourceComponent"/> on an entity on or off
 /// </summary>
 [ByRefEvent]
 public readonly record struct IgnitionEvent(bool Ignite = false);
